api_gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1

PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
